# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auto Mart",
    "description": "Online marketplace for automobiles of diverse makes, model or body type.\n",
    "contact": {
      "email": "aobikobe@gmail.com"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url":"https://quiet-earth-51065.herokuapp.com/api/{version}",
      "variables": {
        "version": {
          "enum": [
             "v1"
          ],
          "default": "v1"
       }
      }
    }
  ],
  "paths": {
    "/car": {
      "get": {
        "tags": [
          "car"
        ],
        "summary": "getCar",
        "description": "Get car",
        "operationId": "Getcar",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status of the car( available is the only option)",
            "required": false,
            "style": "form",
            "example":"available",
            "schema": {
              "enum": [
                "available"
              ]
            }
          },
          {
            "name": "body_type",
            "in": "query",
            "description": "Body type of the car",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min",
            "in": "query",
            "description": "Minimum price of cars",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "max",
            "in": "query",
            "description": "Maximum type of cars",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "manufacturer",
            "in": "query",
            "description": "Manufacturer of cars",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "State of cars",
            "required": false,
            "style": "form",
            "schema": {
              "enum": [
                "new",
                "used"
              ]
            }
          }

        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": [] 
          }
        ]
      },
      "post": {
        "tags": [
          "car"
        ],
        "summary": "create car",
        "description": "Create a new car ad",
        "operationId": "CreateCar",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "state",
                  "price",
                  "model",
                  "bodyType",
                  "manufacturer",
                  "dataFile"
                ],
                "type": "object",
                "properties": {
                  "state": {
                    "enum": ["new","used"],
                    "description": "State of the car ad we want to create"
                  },
                  "price": {
                    "type": "integer",
                    "description": "Price of car ad we want to create"
                  },"model": {
                    "type": "string",
                    "description": "Model of car ad we want to create"
                  },"bodyType": {
                    "type": "string",
                    "description": "Body type of car ad we want to create"
                  },"manufacturer": {
                    "type": "string",
                    "description": "Manufacturer of car ad we want to create"
                  },"dataFile": {
                    "type": "file",
                    "description": "Picture of car ad we want to create"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "car not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": [] 
          }
        ]
      }
      
    },
    "/car/{carId}/{newPrice}": {
      "patch": {
        "tags": [
          "car"
        ],
        "summary": "Update Car Price",
        "description": "Update the price of the car with the specified car id",
        "operationId": "updateCarPrice",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "Id of car",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "newPrice",
            "in": "path",
            "description": "New price of car",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  },
                  "description": "successful operation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": [] 
          }
        ]
      }
    },
...
    "/flag": {
      "post": {
        "tags": [
          "flag"
        ],
        "summary": "postFlag",
        "description": "Report a car ad as fraudulent",
        "operationId": "postFlag",
        "parameters": [],
        "requestBody": {
          "description": "Report an car ad",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "carId",
                  "reason",
                  "description"
                ],
                "type": "object",
                "properties": {
                  "carId": {
                    "type": "integer",
                    "description": "Id of car "
                  },
                  "amount": {
                    "type": "string",
                    "description": "The reason for the complaint"
                  },
                  "description": {
                    "type": "string",
                    "description": "The description of the complaint"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flag"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Flag"
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": [] 
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Car": {
        "title": "Car",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "owner": {
            "type": "integer",
            "format": "int64"
          },
          "price": {
            "type": "integer"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "state": {
            "$ref": "#/components/schemas/State"
          }
          ,
          "manufacturer": {
            "type": "string"
          }
        }
      },
...
      "Flag": {
        "title": "Flag",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Status": {
        "title": "Status",
        "enum": [
          "available",
          "sold"
        ],
        "type": "string",
        "description": "Car Status",
        "example": "available"
      },
      "State": {
        "title": "State",
        "enum": [
          "used",
          "new"
        ],
        "type": "string",
        "description": "State of Vehicle",
        "example": "new"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type":"http",
        "scheme":"bearer",
        "bearerFormat":"JWT",
        "in":"header"
      }
    }
  },
...
}